# [Operator] index_add optimized