--- a/src/flag_gems/ops/index_add.py
+++ b/src/flag_gems/ops/index_add.py
@@ -4,64 +4,221 @@ import os
 from typing import Any, Callable, List, Mapping, Tuple
 
 import torch
-import triton
-import triton.language as tl
 
-from ..runtime import backend
-from ..utils import dim_compress, libentry
-
-device = backend.device_name
+from flag_gems.utils.code_cache import code_cache_dir
+from flag_gems.utils.code_utils import IndentedBuffer
 
 
-def cfggen():
-    block_m = [1, 2, 4]
-    block_n = [128, 1024, 2048, 4096]
-    configs = [
-        triton.Config({"BLOCK_M": m, "BLOCK_N": n}, num_warps=4)
-        for m in block_m
-        for n in block_n
-    ]
-    return configs
+def generate_imports(code: IndentedBuffer) -> IndentedBuffer:
+    code.writeline("import triton")
+    code.writeline("import triton.language as tl")
+    code.writeline("from flag_gems.utils import libentry")
+
+    code.newline()
+    code.newline()
+
+    return code
 
 
-@libentry()
-@triton.autotune(configs=cfggen(), key=["M", "N"])
-@triton.jit
-def index_add_kernel(
-    inp,
-    out,
-    index,
-    src,
-    M,
-    N,
-    alpha,
-    inp_len,
-    BLOCK_M: tl.constexpr,
-    BLOCK_N: tl.constexpr,
-):
-    pid_x = tl.program_id(axis=0)
-    pid_y = tl.program_id(axis=1)
-    rows_offsets = pid_x * BLOCK_M + tl.arange(0, BLOCK_M)[:, None]
-    cols_offsets = pid_y * BLOCK_N + tl.arange(0, BLOCK_N)
+def generate_index_add_kernel(
+    rank: int,
+    kernel_name: str,
+    code: IndentedBuffer,
+) -> IndentedBuffer:
+    # the decorators
+    code.writeline("@libentry()")
+    code.writeline("@triton.jit")
 
-    rows_mask = rows_offsets < M
-    index_mask = cols_offsets < N
-    block_mask = rows_mask and index_mask
+    # signature
+    code.writeline(f"def {kernel_name}(")
+    with code.indent():
+        if rank > 0:
+            code.writeline("index,")
+            code.writeline("src,")
+            code.writeline("out,")
+            code.writeline("N,")
+            code.writeline("inp_numel,")
+            code.writeline("inp_stride_dim,")
+            code.writeline("inp_shape_dim,")
+            code.writeline("src_shape_dim,")
+            code.writeline("delta,")
+            code.writeline("alpha,")
 
-    cur_indices = tl.load(index + cols_offsets, mask=index_mask, other=0)
-    inp_off = rows_offsets * inp_len + cur_indices[None, :]
-    cur_inp = tl.load(inp + inp_off, mask=block_mask, other=0.0)
-    src_off = rows_offsets * N + cols_offsets[None, :]
-    cur_src = tl.load(src + src_off, mask=block_mask, other=0.0)
-    cur_inp += alpha * cur_src
+            stride_args = ", ".join(f"src_stride_{i}: int" for i in range(rank))
+            code.writeline(f"{stride_args}, # stride for src")
 
-    tl.store(out + inp_off, cur_inp, mask=block_mask)
+            shape_args = ", ".join(f"src_shape_{i}: int" for i in range(rank))
+            code.writeline(f"{shape_args}, # shape for src")
+
+            code.writeline("BLOCK_SIZE: tl.constexpr,")
+
+        code.writeline("):")
+
+        # Kernel Code
+        with code.indent():
+            code.writeline("pid = tl.program_id(axis=0)")
+            code.writeline("offsets = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)")
+            code.writeline("mask = offsets < N")
+
+            for i in range(rank - 1, -1, -1):
+                code.writeline(f"src_offset{i} = offsets % src_shape_{i}")
+                code.writeline(f"offsets = offsets // src_shape_{i}")
+            code.newline()
+            comp = [f"src_offset{i} * src_stride_{i}" for i in range(rank)]
+            code.writeline(f"src_offset = {' + '.join(comp)}")
+
+            code.writeline("pre_cal = (inp_stride_dim * src_shape_dim)")
+
+            # index add
+            code.writeline("pre_idx = (src_offset // pre_cal).to(tl.int64)")
+            code.writeline(
+                "dim_idx = (src_offset % pre_cal // inp_stride_dim).to(tl.int64)"
+            )
+            code.writeline(
+                "src_dim_idx = (tl.load(index + dim_idx, mask=mask, other=0)).to(tl.int64)"
+            )
+            code.writeline(
+                'assert src_dim_idx >= 0 and src_dim_idx < inp_shape_dim, "0 <= index < self.size(dim)"'
+            )
+            code.writeline(
+                "input_idx = (src_offset + (delta * pre_idx + src_dim_idx - dim_idx) * inp_stride_dim).to(tl.int64)"
+            )
+
+            code.writeline("input_mask = input_idx < inp_numel")
+            code.writeline(
+                "add_on = tl.load(src + src_offset, mask=mask, other=0) * alpha"
+            )
+            code.writeline(
+                "tl.atomic_add(out + input_idx, add_on, mask=input_mask, sem='relaxed')"
+            )
+            # TODO: tl.atomic_add doesn't support bfloat16! The following method may be unsafe.
+            # code.writeline("cur_out = tl.load(out + input_idx, mask=input_mask)")
+            # code.writeline("tl.store(out + input_idx, cur_out + add_on, mask=input_mask)")
+
+        code.newline()
+        code.newline()
+        return code
+
+
+def parameter_for_wrapper() -> str:
+    # out, index, src, dim, inp_stride_dim, src_shape_dim, delta, N, inp.numel(), alpha
+    parameters: List[str] = []
+    parameters.append("out")
+    parameters.append("index")
+    parameters.append("src")
+    parameters.append("dim")
+    parameters.append("inp_stride_dim")
+    parameters.append("inp_shape_dim")
+    parameters.append("src_shape_dim")
+    parameters.append("delta")
+    parameters.append("N")
+    parameters.append("inp_numel")
+    parameters.append("alpha")
+
+    return ", ".join(parameters)
+
+
+def generate_destination_passing_wrapper(
+    rank: int,
+    wrapper_name: str,
+    kernel_name: str,
+    code: IndentedBuffer,
+) -> IndentedBuffer:
+    parameters: str = parameter_for_wrapper()
+    wrapper_signature: str = f"def {wrapper_name} ({parameters}):"
+    code.writeline(wrapper_signature)
+
+    with code.indent():
+        code.writeline("src_strides = list(src.stride())")
+        code.writeline("src_shapes = list(src.shape)")
+
+        # kernel launch
+        code.writeline("BLOCK_SIZE = 128")  # BLOCK_SIZE setting
+        code.writeline("grid = (triton.cdiv(N, BLOCK_SIZE),)")
+        kernel_launch: str = f"{kernel_name}[grid]("
+        code.writeline(kernel_launch)
+        with code.indent():
+            code.writeline(
+                "index, src, out, N, inp_numel, inp_stride_dim, inp_shape_dim, src_shape_dim, delta, alpha, "
+            )
+            if rank > 0:
+                s = ", ".join(f"src_strides[{i}]" for i in range(rank))
+                code.writeline(f"{s},")
+
+                s = ", ".join(f"src_shapes[{i}]" for i in range(rank))
+                code.writeline(f"{s},")
+            code.writeline("BLOCK_SIZE=BLOCK_SIZE")
+        code.writeline(")")
+        code.writeline("return out")
+
+    return code
+
+
+def generate_code(
+    inputs: Tuple[Any],
+    wrapper_name: str,
+    kernel_name: str,
+    code: IndentedBuffer,
+) -> IndentedBuffer:
+    # inputs: [out, index, src, dim, inp_stride_dim, inp_shape_dim, src_shape_dim, delta, N, inp.numel(), alpha]
+    shape = inputs[2].shape
+    rank = len(shape)
+
+    code = generate_imports(code)
+    code = generate_index_add_kernel(rank, kernel_name, code)
+    code = generate_destination_passing_wrapper(rank, wrapper_name, kernel_name, code)
+    return code
+
+
+class IndexAddFunction:
+    def __init__(self):
+        self.pid = os.getpid()
+        self.overloads: Mapping[str, Callable] = {}
+
+    def __call__(self, *args, **kwargs):
+        key = f"{self.arg_key(*args)}"
+        if key in self.overloads:
+            overload = self.overloads[key]
+        else:
+            code = IndentedBuffer()
+            code = generate_code(
+                args,
+                "_index_add_wrapper",
+                "_index_add_jit_function",
+                code,
+            )
+
+            file_name = f"index_add_rank_{key}_pid_{self.pid}.py"
+
+            with open(code_cache_dir() / file_name, "wt", encoding="utf-8") as f:
+                f.write(code.getvalue())
+
+            # load
+            spec = importlib.util.spec_from_file_location(
+                f"_gen_module_rank_{key}_pid_{self.pid}",
+                f.name,
+            )
+
+            m = importlib.util.module_from_spec(spec)
+            spec.loader.exec_module(m)
+            overload = getattr(m, "_index_add_wrapper")
+            self.overloads[key] = overload
+
+        return overload(*args, **kwargs)
+
+    def arg_key(self, *args):
+        tensors = [item for item in args if torch.is_tensor(item)]
+        max_rank = max(item.ndim for item in tensors)
+        return max_rank
+
+
+_index_add_func = IndexAddFunction()
 
 
 def index_add(inp, dim, index, src, alpha=1):
     logging.debug("GEMS INDEX ADD")
     assert ((0 <= index) * (index < inp.size(dim))).equal(
-        torch.ones(tuple(index.shape), dtype=torch.bool, device=device)
+        torch.ones(tuple(index.shape), dtype=torch.bool, device=inp.device)
     ), "0 <= index < self.size(dim)"
     assert dim >= -inp.ndim and dim < inp.ndim, "Invalid dim"
     assert index.numel() == src.size(