--- a/tests/test_reduction_ops.py
+++ b/tests/test_reduction_ops.py
@@ -700,20 +700,18 @@ def test_accuracy_slice_scatter_with_self_overlapping_input():
     gems_assert_equal(res_out, ref_out)
 
 
-# TODO: failed at (200, 40999, 3)
 @pytest.mark.index_add
 @pytest.mark.parametrize("shape", REDUCTION_SHAPES)
 @pytest.mark.parametrize("dim", DIM_LIST)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+@pytest.mark.parametrize("dtype", [torch.float16, torch.float32])
 def test_accuracy_index_add(shape, dim, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-
+    inp = torch.randn(shape, dtype=dtype, device=flag_gems.device)
     src_shape = list(inp.shape)
     index_max = src_shape[dim]
-    index_len = index_max
-    index = torch.randperm(index_len, device="cuda")
+    index_len = index_max // 2 if index_max >= 2 else 1
+    index = torch.randperm(index_len, device=flag_gems.device)
     src_shape[dim] = index_len
-    src = torch.randn(src_shape, dtype=dtype, device="cuda")
+    src = torch.randn(src_shape, dtype=dtype, device=flag_gems.device)
     alpha = 2
 
     ref_inp = to_reference(inp)