@0x45f 2025-04-01
```
mod = offsets % src_shape_0
src_idx += mod * src_stride_0
offsets = offsets // src_shape_0
mod = offsets % src_shape_1
src_idx += mod * src_stride_1
```
I see code in the generated kernel. Such access seems to be more suitable for column-major tensors? The previous https://github.com/FlagOpen/FlagGems/pull/433 may be helpful.