@GwokHiujin 2025-04-15
Thanks for all the above suggestions; the updated version has been pushed. Any further optimization ideas are welcome :)